updated the data submodule structure and respective tests
